Add github action
When a new PR is opened that modifies the .git-blame-ignore-revs file,
then the new action checks if each line in that file references a
commit in the PR branch. That helps detecting mitakes by which a rebase
or an ammend leads to a new git history that doesn't contain the commit
hash added to .git-blame-ignore-revs anymore.